Added socket world hopper, added xarpus hm def, fixed no outline/width adjustment for socket sote, renamed Socket - DPS to Socket - Damage Counter

+++ plugins/src/main/java/net/runelite/client/plugins/socket/plugins/deathindicators/DeathIndicatorsPlugin.java
@@ -435,7 +435,7 @@ public class DeathIndicatorsPlugin extends Plugin
                     {
                         if(isDamage)
                         {
-                            damage = (int) (Integer.parseInt(text));
+                            damage = Integer.parseInt(text);
                         }
                         else
                         {
